requirement.txt
python-telegram-bot==13.7
redis==6.2.4
requests==2.25.1